Change in diff format: using e.valuelist instead of e.values.
This is to avoid collision with dict.values() in python
and make it possible to use a dict with attribute access
for cleaner code like e.op, e.key, e.valuelist where e
is a diff operation entry.

This commit also enables the use of said dict with attribute access in
the constructed diffs. Some of the previous commits may erroneously
have mentioned this as already in place but it was disabled...