Add a notebook git  diff driver
Adds a diff driver for git, which produces a diff for notebooks only.
This can be used by `git diff`, and relies on filename filtering in
`.gitattributes`to be supplied the right files. The
`.gitattributes`need to be configured for each repo, but the line
included in this commit should be sufficient.