Routines for patching
Routines for patching an object according to the diff, similar to patch
routines in python code. Also includes routines for applying the patch
on the stringified version of the JSON objects, and getting out the
string ranges of the diffs directly. This prevents the need to run
string based diffs on the stringified results.