Implement prettyprinting of notebooks and notebook decisions.
Pretty-printing uses a file-like object to print,
defaulting to sys.stdout. (Currently not using logger).

Call it from merge_notebooks when log level is set to DEBUG
to dump all data in a readable format.

Change behaviour in nbmerge CLI: conflicts are no longer
written in merged notebook metadata, as we're not using
that for the git driver anyway, instead the merge conflicts
are printed to screen (printing needs some more work).