Add placeholder for hidden cell(s)
Replace the ranges of hidden, unchanged cells with "X unchanged cell(s)
hidden". Some tricks were needed, but short summary is:
- Placeholder is put by `::before` pseudoelement on the first visible
cell after a range. If there are no visisble cells after the range, an
`::after` pseudoelement is put on the last visible cell before the
range.
- The cells at which to put the pseudoelements, are marked by a DOM data
attribute. The value of the attribute is the number of cells that were
hidden in the range.
- Some CSS adjustments were needed in order for the pseudoelements to
style well.