Webapp: Add placeholder for external images
When rendering markdown (e.g. unchanged markdown cells), external images
will try to be loaded from the server. Instead of opening up access to
arbitrary files on the server (and figuring out what dir to fix the
server to), we simply show a placeholder image (indicates that at least
the image path is unchanged, even if the image might have changed).