Parse rename metadata CLI parameter
When git detects that it diffs renamed files, it passes two extra
undocumented parameters to the external diff program. One of the
parameters, rename_to, has been supported since 59aef29 (2016-05-03),
but the other has been omitted. This caused (a part of) #497.
This change adds support for the missing CLI parameter.