Fix error with empty base in merge
Git seems to prefer to give a zero-length base file if we have conclicting creation of the same file in two branches, instead of using the explicit missing file (/dev/null or NUL). This adds code to handle this behavior.

We still let nbdime fall over in most other cases where the file is "real", as empty .ipynb files are invalid JSON.