Update typescript (#641)
* Update typescript

I am working on another pr which uses newer typescript features, so I thought I would update it first.
This pr updates to typescript 4.9 and fixes the resulting errors. It also
sets up a stricter tsconfig_base which is used by all tsconfig files and
fixes the resulting errors.

* add back funciton that maybe needed externally

* fix test typescript errors

* Cleanup TS upgrade

Reverts parameter renames, and removes tsconfig settings with the default values. Removes as error expected directives, and fixes typings in another way.

* Enforce explicit type-only imports

This forces type-only imports to be imported with `import type`

* CI py 3.6 -> 3.11

* Fix cookie secret

* Remove notebook tests

With jupyter_server 2.0, we cannot support launching with notebook. The tests can potentially be reenabled with notebook 7.0, as it is based on jupyter_server.

* Add token auth fixture

When there is a token, xsrf cookies/headers are ignored, so using it makes tests easier.

Co-authored-by: Vidar Tonaas Fauske <510760+vidartf@users.noreply.github.com>
Co-authored-by: Vidar Tonaas Fauske <vidartf@gmail.com>