Change header for unchanged notebook meta (#736)
* Change header for unchanged notebook meta

We previously changed to always showing the metadata diff in notebook merges, in case the user needs to make manual edits to it even if there are no changes in the merge itself. That change did not take into account that the display behavior here made some assumptions about only being shown when there was a change. Here we make the title conditional on whether there are changes or not, and make sure not to collapse similar changes in the case where there are none.

* Update Playwright Snapshots

* Update Playwright Snapshots

* Disable animations for metadata drawer expand

* Update Playwright Snapshots

* Update Playwright Snapshots

* fix string for translation

Co-authored-by: Michał Krassowski <5832902+krassowski@users.noreply.github.com>

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Michał Krassowski <5832902+krassowski@users.noreply.github.com>